requirements.txt
scapy~=2.5.0
pyyaml